requirements.txt
torch
torchvision
git+https://github.com/huggingface/transformers
accelerate
qwen-vl-utils
fastapi
uvicorn
matplotlib
moondream
requests
Pillow
